fix: avoid importing image manipulator in the popup script
